test(IAM Access Groups): additional fixes to reduce integration test failures (#16)
* chore: fix linter issues in int test

+++ test/integration/iam-access-groups.v2.test.js
@@ -2,17 +2,17 @@
 /**
  * (C) Copyright IBM Corp. 2020.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
  */
 
 'use strict';
@@ -40,7 +40,7 @@ describe('IamAccessGroupsV2_integration', () => {
   const testGroupDescription = 'This group is used for integration test purposes. It can be deleted at any time.';
   let testGroupETag;
   let testGroupId;
-  const testUserId = 'IBMid-1234';
+  const testUserId = 'IBMid-' + Math.floor(Math.random() * 100000);
   const userType = 'user';
   let testClaimRuleId;
   let testClaimRuleETag;
@@ -596,22 +596,28 @@ describe('IamAccessGroupsV2_integration', () => {
     // Iterate across the groups
     let group;
     for (group of result.groups) {
-      // Force delete each test group
+      // Force delete the test group (or any test groups older than 5 minutes)
       if (group.name === testGroupName) {
-        const params = {
-          accessGroupId: group.id,
-          force: true,
-        };
+        const createdAt = Date.parse(group.created_at);
+        const FIVE_MINUTES = 5 * 60 * 1000;
+        const fiveMinutesAgo = Date.now() - FIVE_MINUTES;
 
-        let response;
-        try {
-          response = await service.deleteAccessGroup(params);
-        } catch (err) {
-          done(err);
+        if (group.id === testGroupId || createdAt < fiveMinutesAgo) {
+          const params = {
+            accessGroupId: group.id,
+            force: true,
+          };
+
+          let response;
+          try {
+            response = await service.deleteAccessGroup(params);
+          } catch (err) {
+            done(err);
+          }
+
+          expect(response).toBeDefined();
+          expect(response.status).toEqual(204);
         }
-
-        expect(response).toBeDefined();
-        expect(response.status).toEqual(204);
       }
     }
